backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yfinance==0.2.36
pandas==2.2.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
feedparser==6.0.11
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
pytest==8.0.0
pytest-asyncio==0.23.3
